add scripts to build and push gems
